requirements.txt
requests>=2.31.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0